clean: ManuScript to worker
