+++ crates/audio_generator/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "script_audio_generator"
+name = "audio_generator"
 version = "0.1.0"
 edition = "2021"
