+++ crates/audio_generator/src/lib.rs
