+++ crates/audio_generator/src/voicevox/client.rs
@@ -1,5 +1,18 @@
 use super::speaker::VoiceVoxSpeaker;
+use crate::AudioGenerator;
+use anyhow::Result;
+use async_trait::async_trait;
 use serde_json::Value;
+
+#[async_trait]
+impl AudioGenerator for VoiceVoxClient {
+    async fn generate(&self, speaker_id: &str, text: &str) -> Result<Vec<u8>> {
+        let speaker = serde_json::from_str(speaker_id)?;
+        let query = self.query(text, &speaker).await?;
+        let audio = self.synthesis(query, &speaker).await?;
+        Ok(audio)
+    }
+}
 
 #[derive(Debug, Clone)]
 pub struct VoiceVoxClient {
