+++ crates/audio_generator/src/voicevox/mod.rs
@@ -1,0 +1,2 @@
+pub mod client;
+mod speaker;
