+++ crates/worker/Cargo.toml
@@ -16,7 +16,7 @@ path = "src/bin/cli.rs"
 readable_text = { path = "../readable_text" }
 script_http_client = { path = "../script_http_client" }
 script_runtime = { path = "../script_runtime" }
-script_audio_generator = { path = "../script_audio_generator" }
+audio_generator = { path = "../audio_generator" }
 repos = { path = "../repos" }
 anyhow = "1.0.86"
 axum = "0.7.5"
